Applying refactorings
