+++ util/mutant/__init__.py
@@ -1,0 +1,21 @@
+from .mutant import Mutant
+from .c_mutator import CMutator
+from .mutation_operator import (
+    MutationOperator,
+    ArithmeticOperatorReplacement,
+    RelationalOperatorReplacement,
+    LogicalConnectorReplacement,
+    ConstantReplacement,
+    ReturnValueReplacement,
+)
+
+__all__ = [
+    "CMutator",
+    "Mutant",
+    "MutationOperator",
+    "ArithmeticOperatorReplacement",
+    "RelationalOperatorReplacement",
+    "LogicalConnectorReplacement",
+    "ConstantReplacement",
+    "ReturnValueReplacement",
+]
